done requested changes
